refactor: 컨트롤러 메서드 반환값 수정
- 통일이 안된 메서드 리턴 방식을 ResponseEntity<?>로 리턴받도록 수정
